requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pillow==10.4.0
numpy==1.26.4
tensorflow==2.17.0
requests==2.32.3
python-multipart==0.0.9
